Move `backoff` to the internal/backoff pkg for sharing between slack pkg and slacksocketmode pkg
